+++ include/CacheDisk.h
@@ -50,18 +50,18 @@ namespace openshot {
 	class CacheDisk : public CacheBase {
 	private:
 		QDir path; ///< This is the folder path of the cache directory
-		map<long int, long int> frames;	///< This map holds the frame number and Frame objects
-		deque<long int> frame_numbers;	///< This queue holds a sequential list of cached Frame numbers
+		map<int64_t, int64_t> frames;	///< This map holds the frame number and Frame objects
+		deque<int64_t> frame_numbers;	///< This queue holds a sequential list of cached Frame numbers
 		string image_format;
 		float image_quality;
 		float image_scale;
 
-		long long int frame_size_bytes; ///< The size of the cached frame in bytes
+		int64_t frame_size_bytes; ///< The size of the cached frame in bytes
 		bool needs_range_processing; ///< Something has changed, and the range data needs to be re-calculated
 		string json_ranges; ///< JSON ranges of frame numbers
-		vector<long int> ordered_frame_numbers; ///< Ordered list of frame numbers used by cache
-		map<long int, long int> frame_ranges;	///< This map holds the ranges of frames, useful for quickly displaying the contents of the cache
-		long int range_version; ///< The version of the JSON range data (incremented with each change)
+		vector<int64_t> ordered_frame_numbers; ///< Ordered list of frame numbers used by cache
+		map<int64_t, int64_t> frame_ranges;	///< This map holds the ranges of frames, useful for quickly displaying the contents of the cache
+		int64_t range_version; ///< The version of the JSON range data (incremented with each change)
 
 		/// Clean up cached frames that exceed the max number of bytes
 		void CleanUp();
@@ -86,7 +86,7 @@ namespace openshot {
 		/// @param quality The quality of the image (1.0=highest quality/slowest speed, 0.0=worst quality/fastest speed)
 		/// @param scale The scale factor for the preview images (1.0 = original size, 0.5=half size, 0.25=quarter size, etc...)
 		/// @param max_bytes The maximum bytes to allow in the cache. Once exceeded, the cache will purge the oldest frames.
-		CacheDisk(string cache_path, string format, float quality, float scale, long long int max_bytes);
+		CacheDisk(string cache_path, string format, float quality, float scale, int64_t max_bytes);
 
 		// Default destructor
 		~CacheDisk();
@@ -99,30 +99,30 @@ namespace openshot {
 		void Clear();
 
 		/// Count the frames in the queue
-		long int Count();
+		int64_t Count();
 
 		/// @brief Get a frame from the cache
 		/// @param frame_number The frame number of the cached frame
-		std::shared_ptr<Frame> GetFrame(long int frame_number);
+		std::shared_ptr<Frame> GetFrame(int64_t frame_number);
 
 		/// Gets the maximum bytes value
-		long long int GetBytes();
+		int64_t GetBytes();
 
 		/// Get the smallest frame number
 		std::shared_ptr<Frame> GetSmallestFrame();
 
 		/// @brief Move frame to front of queue (so it lasts longer)
 		/// @param frame_number The frame number of the cached frame
-		void MoveToFront(long int frame_number);
+		void MoveToFront(int64_t frame_number);
 
 		/// @brief Remove a specific frame
 		/// @param frame_number The frame number of the cached frame
-		void Remove(long int frame_number);
+		void Remove(int64_t frame_number);
 
 		/// @brief Remove a range of frames
 		/// @param start_frame_number The starting frame number of the cached frame
 		/// @param end_frame_number The ending frame number of the cached frame
-		void Remove(long int start_frame_number, long int end_frame_number);
+		void Remove(int64_t start_frame_number, int64_t end_frame_number);
 
 		/// Get and Set JSON methods
 		string Json(); ///< Generate JSON string of this object
